EntregaFinal

+++ Proyecto.Ecommerce/Proyecto.Ecommerce.Aplicacion/Dtos/OrdenDto.cs
@@ -12,8 +12,8 @@ namespace Proyecto.Ecommerce.Aplicacion.Dtos
         public string NombreCliente { get; set; }
         public string MetodoEntregaId { get; set; }
         public string NombreMetodoEntrega { get; set; }
-        public long SubTotal { get; set; }
+        public float SubTotal { get; set; }
         public int Iva { get; set; }
-        public long Total { get; set; }
+        public float Total { get; set; }
     }
 }
