+++ Proyecto.Ecommerce/Proyecto.Ecommerce.Dominio/Entidades/Orden.cs
@@ -10,8 +10,8 @@ namespace Proyecto.Ecommerce.Dominio.Entidades
         public List<OrdenItems>? OrdenItems { get; set; }
         public MetodoEntrega? MetodoEntrega { get; set; }
         public string MetodoEntregaId { get; set; }
-        public long SubTotal { get; set; }
+        public float SubTotal { get; set; }
         public int Iva { get; set; }
-        public long Total { get; set; }
+        public float Total { get; set; }
     }
 }
